Create Moon-phase.py - Trianglium (#44)
# Moon Phase 
### A Simple program to calculate the current (approximate) Moon Phase.  

## Requirements
This program uses a module in python's standard library, 'datetime'

## Inspiration 
['Calculate Moon Phase' by Subsystems.us](https://www.subsystems.us/uploads/9/8/9/4/98948044/moonphase.pdf)

### Status
This program runs completely fine, 'as is'. Though, **I'm open to any suggestions on how to make the moon age and moon phase names, more accurate**. There was mixed information online and the phases mentioned in the 'inspiration' article, were opposite of the actual moon phases. 


***hacktoberfest2021***