Virus
It is a virus program which duplicates itself into all the files with extension .py , .pyw ,.cpp , .c containing in the same folder as this file. For its implementation sys and glob libraries used to open the files and reading their extensions respectively . A loop is used , to check whether the file is corrupted or not with the checking condition "#START OF VIRUS#" , like if start of virus not present in the file then it will be corrupted.