Virus (#57)
* Spam Bot

It will spam anything u save in spam.txt, on WhatsApp Instagram etc

* Snake Game

This is the retro snake game I made , it uses turtle, randrange, square and vector library , for the implementation I made three functions change, inside and move responsible for changing direction, to check if it is within boundaries and to move respectively, and onkey is also used to define to change the snake's direction to left right up and down

* Rock Paper Scissors Game

A simple rock paper scissor game program , random library used to use randInt, the user inputs what he has to choose between rock paper scissor which are labelled as 1, 2,0 respectively and random integer generated for the computer and if u win lose or match drawn , the message is displayed accordingly.

* Delete rock_paper_scissors.py

* Virus

It is a virus program which duplicates itself into all the files with extension .py , .pyw ,.cpp , .c containing in the same folder as this file. For its implementation sys and glob libraries used to open the files and reading their extensions respectively . A loop is used , to check whether the file is corrupted or not with the checking condition "#START OF VIRUS#" , like if start of virus not present in the file then it will be corrupted.